# indexcards/api/routes/openapi/openapi.json
{
  "openapi": "3.1.1",
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "info": {
    "title": "IndexCards API",
    "version": "1.2.0",
    "description": "Tabroom.com data & operational API",
    "termsOfService": "https://www.speechanddebate.org/terms-conditions/",
    "license": {
      "name": "Copyright 2014-2021, National Speech & Debate Association"
    }
  },
  "security": [],
...
  "paths": {
...
    "/rest/tourns/{tournId}/invite": {
      "get": {
        "summary": "Get Tournament Invite",
        "operationId": "getTournInvite",
        "description": "Retrieve a public invite for a specific tournament, including pages, files, events, and contacts.",
        "tags": [
          "Tournaments",
          "test"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/tournId"
          }
        ],
        "responses": {
          "200": {
            "description": "Public facing page data for a given tournament",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournInviteSchema"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TournContactSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "first": {
            "type": "string"
          },
          "middle": {
            "type": [
              "string",
              "null"
            ]
          },
          "last": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          }
        },
        "required": [
          "id",
          "first",
          "middle",
          "last",
          "email"
        ],
        "additionalProperties": false
      },
...
      "ProblemSchema": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "A URI reference that identifies the problem type."
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "The HTTP status code."
          },
          "detail": {
            "description": "Human-readable explanation of the error.",
            "type": "string"
          },
          "instance": {
            "description": "A URI reference to the specific occurrence.",
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "additionalProperties": false
      },
...
      "TournInviteSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "name": {
            "type": "string",
            "maxLength": 63
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 31
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6
              },
              {
                "type": "null"
              }
            ]
          },
          "country": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4
              },
              {
                "type": "null"
              }
            ]
          },
          "tz": {
            "type": "string",
            "maxLength": 31
          },
          "webname": {
            "type": "string"
          },
          "hidden": {
            "type": "boolean"
          },
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "regStart": {
            "type": "string"
          },
          "regEnd": {
            "type": "string"
          },
          "Webpages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebpageSchemaOutput"
            }
          },
          "Files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileSchemaOutput"
            }
          },
          "Events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventOutput"
            }
          },
          "Contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TournContactSchema"
            }
          }
        },
        "required": [
          "id",
          "name",
          "city",
          "state",
          "country",
          "tz",
          "webname",
          "hidden",
          "start",
          "end",
          "regStart",
          "regEnd",
          "Webpages",
          "Files",
          "Events",
          "Contacts"
        ],
        "additionalProperties": false
      },
...
      "EventOutput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "abbr": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "fee": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "debate",
              "speech",
              "mock_trial",
              "congress",
              "wsdc",
              "wudc",
              "attendee",
              "academic"
            ]
          },
          "categoryId": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "settings": {
            "type": "object",
            "properties": {
              "flight_offset": {
                "anyOf": [
                  {
                    "type": "integer",
                    "minimum": -9007199254740991,
                    "maximum": 9007199254740991
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "online_mode": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "online_ballots": {
                "type": "boolean"
              }
            },
            "required": [
              "flight_offset",
              "online_mode",
              "online_ballots"
            ],
            "additionalProperties": false
          },
          "metadata": {
            "type": "object",
            "properties": {},
            "additionalProperties": false
          },
          "nsdaCategoryId": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "abbr",
          "name",
          "fee",
          "type",
          "categoryId",
          "settings",
          "metadata",
          "nsdaCategoryId"
        ],
        "additionalProperties": false
      },
      "FileSchemaOutput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "description": "The unique identifier for the file"
          },
          "tag": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": [
              "string",
              "null"
            ]
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "filename": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ]
          },
          "published": {
            "type": "boolean"
          },
          "page_order": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "uploaded": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
              },
              {
                "type": "null"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": [
          "id",
          "tag",
          "type",
          "label",
          "filename",
          "published",
          "page_order",
          "uploaded",
          "timestamp"
        ],
        "additionalProperties": false
      },
...
      "WebpageSchemaOutput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 63
              },
              {
                "type": "null"
              }
            ]
          },
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "published": {
            "type": "boolean"
          },
          "sitewide": {
            "type": "boolean"
          },
          "special": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 15
              },
              {
                "type": "null"
              }
            ]
          },
          "slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 63
              },
              {
                "type": "null"
              }
            ]
          },
          "pageOrder": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "parentId": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "published",
          "sitewide",
          "special",
          "slug",
          "pageOrder",
          "parentId",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
...
      "ErrorResponse": {
        "description": "Unexpected Error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemSchema"
            },
            "examples": {
              "error": {
                "summary": "500 Internal Server Error",
                "value": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "An unexpected error occurred.",
                  "instance": "/api/v1/example"
                }
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "NotFound - the requested resource was not found or you do not have access.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemSchema"
            },
            "examples": {
              "notfound": {
                "summary": "404 Not Found",
                "value": {
                  "type": "about:blank",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource foo with ID bar was not found.",
                  "instance": "/api/v1/example"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - authentication failed or was not provided.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemSchema"
            },
            "examples": {
              "unauthorized": {
                "summary": "401 Unauthorized",
                "value": {
                  "type": "about:blank",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "You are not authorized to access this resource."
                }
              }
            }
          }
        }
      }
    },
    "parameters": {
      "tournId": {
        "name": "tournId",
        "in": "path",
        "description": "The ID or webname of the tournament to retrieve.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "securitySchemes": {
      "extApiKey": {
        "type": "http",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "TabroomToken",
        "description": "send the session token as a cookie. For direct API access, we recommend using bearerAuth."
      }
    }
  }
}